move subscriptions client to its own file
